Split tmux control byte helpers
